BACKEND/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
efficientnet_pytorch
albumentations>=1.3.0
opencv-python
scikit-learn
pillow
numpy
matplotlib
tqdm
